refactor: split mock api by api category
