+++ src/mocks/apis/room/index.ts
@@ -1,0 +1,13 @@
+import { rest } from "msw"
+import { baseUrl } from "@/mocks/config"
+
+export default [
+  rest.get(`${baseUrl}/room`, (req, res, ctx) => {
+    return res(
+      ctx.status(200),
+      ctx.json({
+        roomId: "1",
+      }),
+    )
+  }),
+]
