+++ src/mocks/config.ts
@@ -1,0 +1,1 @@
+export const baseUrl = import.meta.env.VITE_API_BASE_URL
